+++ resources/views/users/index.blade.php
@@ -44,14 +44,16 @@
                                     @endcan
                                 </td>
                                 <td>
-                                    {{ Form::open([
-                                            'route' => ['users.destroy',$user->id], 'method'=>'DELETE'
-                                        ]) 
-                                    }}
+                                    @can('users.destroy')    
+                                        {{ Form::open([
+                                                'route' => ['users.destroy',$user->id], 'method'=>'DELETE'
+                                            ]) 
+                                        }}
 
-                                        <button class="btn btn-sm btn-danger">Eliminar</button>
+                                            <button class="btn btn-sm btn-danger">Eliminar</button>
 
-                                    {{ Form::close() }}
+                                        {{ Form::close() }}
+                                    @endcan
                                 </td>
                             </tr>
                             @endforeach
